Changed memory types to big ints to hold more than 2GB of RAM.

+++ ufdl-core-app/src/ufdl/core_app/models/nodes/_Node.py
@@ -34,10 +34,10 @@ class Node(DeleteOnNoRemainingReferencesOnlyModel):
                                             null=True)
 
     # The amount of GPU memory available on the node, in MB
-    gpu_mem = models.PositiveIntegerField(null=True)
+    gpu_mem = models.BigIntegerField(null=True)
 
     # The amount of CPU memory available on the node, in MB
-    cpu_mem = models.PositiveIntegerField()
+    cpu_mem = models.BigIntegerField()
 
     # The timestamp when the node last made contact
     last_seen = models.DateTimeField(null=True, default=None)
